@1.2::IServiceManager: Implement new onClients API
This now reflects state changes from no clients -> clients in addition
to from clients -> no clients. For full details, see the system/libhidl
CL where this introduced.

The most important thing here is that when a HAL is retrieved (::get),
onClient(true) status will be immediately updated so that no race is
possible.

Bug: 117779690
Test: hidl_test
Test: boot Pixel 2 and manually test this API
Change-Id: Ifc4dfa79233c8f5b2aff490e78214a1788bf6e6c