lazy hals: fix onClient(true) double-send.
Before, if guaranteeClient was called, and mHasClients was true
(meaning that we should have a client and there is a record of having
a client), then an onClient(true) notification would be called.
However, since mHasClients is true, we must have already sent a
notification out. A CHECK for this situation was also added.

Test: hwservicemanager_test
    HidlServiceLazyTest AcquisitionAfterGuarantee (now passes)
Test: internal 'lazy_hidl_test' fuzzing health storage HAL, health
    storage HAL no longer reports double-acquisition of clients.

Change-Id: I18489c1b1d085028ce39163035a06ca04fd9f10d